requirements.txt
pdfminer.six
pyyaml
rapidfuzz
whisper
torch
